Issue #380 - Alpaca Crypto Data API (v1-beta1) implementation (RESTFul API part).

+++ Alpaca.Markets/Interfaces/IBar.cs
@@ -49,7 +49,7 @@ namespace Alpaca.Markets
         /// Gets bar trading volume.
         /// </summary>
         [UsedImplicitly]
-        UInt64 Volume { get; }
+        Decimal Volume { get; }
         
         /// <summary>
         /// Gets bar volume weighted average price.
